+++ Extensions/ComparableExtensions.cs
@@ -14,7 +14,6 @@
             /// 当 <paramref name="value"/> 小于下限 <paramref name="min"/> 时，会被重新赋值为 <paramref name="min"/>，否则不做改变
             /// </para>
             /// </summary>
-            /// <param name="value"></param>
             /// <param name="min">下限值</param>
             /// <returns>
             /// 当 <paramref name="value"/> 小于下限 <paramref name="min"/> 时，返回 <paramref
@@ -37,7 +36,6 @@
             /// <paramref name="min"/>； 其余情况不做改变
             /// </para>
             /// </summary>
-            /// <param name="value"></param>
             /// <param name="max">上限值</param>
             /// <param name="min">下限值</param>
             /// <returns>
@@ -56,7 +54,6 @@
             /// 当 <paramref name="value"/> 大于上限 <paramref name="max"/> 时，会被重新赋值为 <paramref name="max"/>，否则不做改变
             /// </para>
             /// </summary>
-            /// <param name="value"></param>
             /// <param name="max">上限值</param>
             /// <returns>
             /// 当 <paramref name="value"/> 大于上限 <paramref name="max"/> 时，返回 <paramref